requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.4
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.6.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.6.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
